refactor(visual): pin the capture landscape where screens are opened
openEntry is what opens a screen — through the seam for a window, over
its host for a modal, in a window of its own for a component — so the
landscape they are opened in belongs there rather than split across the
runner. It also catches the kind the seed alone could not: a component's
host is a bare window with no TopLevelWindow to carry the pin, so it
takes orientationModes directly.

Refs WB-273
