+++ walta-app/app/spec/visual/openEntry.js
@@ -23,6 +23,21 @@
 var { wrapViewInWindow, windowOpenTest, closeWindow } = require("spec/util/TestUtils");
 var { makeTestServices } = require("spec/fixtures/Services_fixture");
 
+// Every screen is opened in one landscape. iOS re-resolves a window's
+// orientation while the device is flat, as a simulator always is, so screens
+// otherwise settle in either — which turns the captured frame the other way up
+// and mirrors the safe-area insets, so the notch changes sides between runs and
+// no baseline holds. Asking the device which way it went doesn't work: a window
+// reports the interface orientation, which stays put while it renders rotated.
+var CAPTURE_LANDSCAPE = Ti.UI.LANDSCAPE_RIGHT;
+
+// TopLevelWindow pins each window it opens to the landscape already in use, so
+// seeding that pins the screen about to open. It has to be re-seeded per screen:
+// each window writes back the orientation it read once laid out.
+function pinToCaptureLandscape() {
+	Alloy.Globals.lastLandscapeOrientation = CAPTURE_LANDSCAPE;
+}
+
 // A screen only one platform ever instantiates (the iOS inline date picker
 // crashes Titanium's Android TextInputLayout, so Android uses the native dialog).
 function runsHere(entry) {
@@ -43,6 +58,7 @@ function screenOf(entry) {
 async function openComponent(entry) {
 	var ctl = Alloy.createController(screenOf(entry), entry.args());
 	var win = wrapViewInWindow(ctl.getView());
+	win.orientationModes = [CAPTURE_LANDSCAPE];
 	await windowOpenTest(win);
 	return {
 		view: win,
@@ -83,6 +99,7 @@ async function withEntryState(entry, opened) {
 }
 
 async function openEntry(entry, entries) {
+	pinToCaptureLandscape();
 	// A component has no seam: it is built directly and hosted in a window.
 	if (entry.wrap) { return withEntryState(entry, await openComponent(entry)); }
 
@@ -104,4 +121,5 @@ async function openEntry(entry, entries) {
 }
 
 exports.openEntry = openEntry;
+exports.CAPTURE_LANDSCAPE = CAPTURE_LANDSCAPE;
 exports.runsHere = runsHere;
